src: kw_manage_contacts: add show groups and group contacts infos
To support the new feature `kw manage-contacts`, it should be able to
display the groups infos after they are created or the contacts infos
from an specific group. In this sense, add the new functionality
`--grops-show` wich can recieve an specific group infos and it's
contacts or all the groups infos in case no parameter is passed.

Signed-off-by: JGBSouza <joaosouzaaa12@gmail.com>